+++ class13/Hwangjunhong/ArchitectureStudy/app/build.gradle
@@ -1,8 +1,6 @@
 apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
-apply plugin: 'kotlin-kapt'
-apply plugin: 'dagger.hilt.android.plugin'
+apply from: '../config.gradle'
+
 
 android {
     compileSdkVersion 29
@@ -15,8 +13,6 @@ android {
         versionCode 1
         versionName "1.0"
 
-        buildConfigField "String", "NAVER_CLIENT_ID", '\"baPVy6Scd6jvsqAsrrHu\"'
-        buildConfigField "String", "NAVER_CLIENT_SECRET", '\"lCfi3Abd26\"'
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
 
@@ -27,15 +23,6 @@ android {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
-    }
-
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-
-    kotlinOptions {
-        jvmTarget = "1.8"
     }
 
     buildFeatures {
@@ -51,12 +38,7 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.2.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.0.1'
 
-    // Hilt
-    implementation "com.google.dagger:hilt-android:2.28-alpha"
-    kapt "com.google.dagger:hilt-android-compiler:2.28-alpha"
-    implementation 'androidx.lifecycle:lifecycle-viewmodel-savedstate:2.2.0'
-    implementation 'androidx.hilt:hilt-lifecycle-viewmodel:1.0.0-alpha01'
-    kapt 'androidx.hilt:hilt-compiler:1.0.0-alpha01'
+    implementation project(path: ':data')
 
     // Rx
     def rxjava_version = '2.2.19'
@@ -77,23 +59,8 @@ dependencies {
     implementation "androidx.activity:activity-ktx:1.1.0"
     implementation "androidx.fragment:fragment-ktx:1.2.5"
 
-    //Room
-    def room_version = "2.2.5"
-    implementation "androidx.room:room-runtime:$room_version"
-    kapt "androidx.room:room-compiler:$room_version"
-    implementation "androidx.room:room-ktx:$room_version"
-
     //RecyclerView
     implementation "androidx.recyclerview:recyclerview:1.2.0-alpha05"
-
-    //Retrofit
-    implementation 'com.google.code.gson:gson:2.8.6'
-    implementation 'com.squareup.retrofit2:retrofit:2.9.0'
-    implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
-
-    //Okhttp
-    implementation 'com.squareup.okhttp3:logging-interceptor:4.8.1'
-    implementation 'com.squareup.okhttp3:okhttp:4.8.1'
 
     //Glide
     implementation 'com.github.bumptech.glide:glide:4.11.0'
