--- class13/Hwangjunhong/ArchitectureStudy/app/src/main/java/com/hong/architecturestudy/MainModule.kt
@@ -1,39 +1,0 @@
-package com.hong.architecturestudy
-
-import com.hong.architecturestudy.data.repository.RepositoryDataSource
-import com.hong.architecturestudy.data.repository.RepositoryDataSourceImpl
-import com.hong.architecturestudy.data.source.local.LocalDataSource
-import com.hong.architecturestudy.data.source.local.LocalDataSourceImpl
-import com.hong.architecturestudy.data.source.remote.RemoteDataSource
-import com.hong.architecturestudy.data.source.remote.RemoteDataSourceImpl
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ApplicationComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(ApplicationComponent::class)
-abstract class RepositoryModule {
-
-    @Binds
-    @Singleton
-    abstract fun bindRepository(repositoryDataSourceImpl: RepositoryDataSourceImpl): RepositoryDataSource
-}
-
-@Module
-@InstallIn(ApplicationComponent::class)
-abstract class LocalDataSourceModule {
-    @Binds
-    @Singleton
-    abstract fun bindLocalDataSource(localDataSourceImpl: LocalDataSourceImpl): LocalDataSource
-}
-
-@Module
-@InstallIn(ApplicationComponent::class)
-abstract class RemoteDataSourceModule {
-
-    @Binds
-    @Singleton
-    abstract fun bindRemoteDataSourceImpl(remoteDataSourceImpl: RemoteDataSourceImpl): RemoteDataSource
-}
