+++ class13/Hwangjunhong/ArchitectureStudy/app/src/main/java/com/hong/architecturestudy/MyApplication.kt
@@ -1,20 +1,7 @@
 package com.hong.architecturestudy
 
 import android.app.Application
-import android.content.Context
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class MyApplication : Application() {
-
-    init {
-        instance = this
-    }
-
-    companion object {
-        private var instance: MyApplication? = null
-        fun applicationContext(): Context {
-            return instance!!.applicationContext
-        }
-    }
-} 
+class MyApplication : Application()
