--- class13/Hwangjunhong/ArchitectureStudy/app/src/main/java/com/hong/architecturestudy/constants/Constants.kt
@@ -1,6 +1,0 @@
-package com.hong.architecturestudy.constants
-
-object Constants {
-    const val BASE_URL = "https://openapi.naver.com/"
-    const val DEFAULT_TIME_OUT: Long = 15
-}
