--- class13/Hwangjunhong/ArchitectureStudy/app/src/main/java/com/hong/architecturestudy/data/source/local/LocalDataSource.kt
@@ -1,8 +1,0 @@
-package com.hong.architecturestudy.data.source.local
-
-import com.hong.architecturestudy.data.source.local.entity.MovieInfo
-
-interface LocalDataSource {
-    fun saveRecentSearchQuery(keyword: String)
-    fun loadRecentSearchQuery(): List<MovieInfo>
-}
