--- class13/Hwangjunhong/ArchitectureStudy/app/src/main/java/com/hong/architecturestudy/data/source/local/entity/MovieInfo.kt
@@ -1,13 +1,0 @@
-package com.hong.architecturestudy.data.source.local.entity
-
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "movie")
-data class MovieInfo(
-    @PrimaryKey(autoGenerate = true)
-    val id: Long? = null,
-    @ColumnInfo(name = "movietitle")
-    var movieTitle: String = ""
-)
