--- class13/Hwangjunhong/ArchitectureStudy/app/src/main/java/com/hong/architecturestudy/data/source/remote/RemoteDataSource.kt
@@ -1,12 +1,0 @@
-package com.hong.architecturestudy.data.source.remote
-
-import com.hong.architecturestudy.data.model.MovieData
-
-interface RemoteDataSource {
-
-    fun getMovieList(
-        query: String,
-        onSuccess: (List<MovieData>) -> Unit,
-        onFailure: (Throwable) -> Unit
-    )
-}
