+++ class13/Hwangjunhong/ArchitectureStudy/app/src/main/java/com/hong/architecturestudy/ui/main/holder/MovieViewHolder.kt
@@ -1,8 +1,8 @@
 package com.hong.architecturestudy.ui.main.holder
 
 import androidx.recyclerview.widget.RecyclerView
-import com.hong.architecturestudy.data.model.MovieData
 import com.hong.architecturestudy.databinding.ItemMovieListBinding
+import com.hong.data.model.MovieData
 
 class MovieViewHolder(private val binding: ItemMovieListBinding) : RecyclerView.ViewHolder(binding.root) {
