+++ class13/Hwangjunhong/ArchitectureStudy/app/src/main/java/com/hong/architecturestudy/ui/moviedialog/holder/MovieSearchListViewHolder.kt
@@ -1,9 +1,9 @@
 package com.hong.architecturestudy.ui.moviedialog.holder
 
 import androidx.recyclerview.widget.RecyclerView
-import com.hong.architecturestudy.data.source.local.entity.MovieInfo
 import com.hong.architecturestudy.databinding.ItemMovieSearchListBinding
 import com.hong.architecturestudy.ui.main.MainViewModel
+import com.hong.data.source.local.entity.MovieInfo
 
 class MovieSearchListViewHolder(private val binding: ItemMovieSearchListBinding) : RecyclerView.ViewHolder(binding.root) {
