+++ class13/Hwangjunhong/ArchitectureStudy/data/.gitignore
@@ -1,0 +1,1 @@
+/build
