+++ class13/Hwangjunhong/ArchitectureStudy/data/src/main/AndroidManifest.xml
@@ -1,0 +1,5 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.hong.data">
+
+    /
+</manifest>
