+++ class13/Hwangjunhong/ArchitectureStudy/data/src/main/java/com/hong/data/constants/Constants.kt
@@ -1,0 +1,6 @@
+package com.hong.data.constants
+
+object Constants {
+    const val BASE_URL = "https://openapi.naver.com/"
+    const val DEFAULT_TIME_OUT: Long = 15
+}
