+++ class13/Hwangjunhong/ArchitectureStudy/data/src/main/java/com/hong/data/source/local/LocalDataSource.kt
@@ -1,0 +1,8 @@
+package com.hong.data.source.local
+
+import com.hong.data.source.local.entity.MovieInfo
+
+interface LocalDataSource {
+    fun saveRecentSearchQuery(keyword: String)
+    fun loadRecentSearchQuery(): List<MovieInfo>
+}
