+++ class13/Hwangjunhong/ArchitectureStudy/data/src/main/java/com/hong/data/source/remote/RemoteDataSource.kt
@@ -1,0 +1,12 @@
+package com.hong.architecturestudy.data.source.remote
+
+import com.hong.data.model.MovieData
+
+interface RemoteDataSource {
+
+    fun getMovieList(
+        query: String,
+        onSuccess: (List<MovieData>) -> Unit,
+        onFailure: (Throwable) -> Unit
+    )
+}
