+++ class13/Hwangjunhong/ArchitectureStudy/settings.gradle
@@ -1,2 +1,3 @@
+include ':data'
 include ':app'
 rootProject.name = "ArchitectureStudy"
